+++ app/services/students_service.rb
@@ -25,22 +25,27 @@ class StudentsService
     end
 
     def get_my_interested_meetings
-      query=AvailabilityTutor.joins(:interested).joins(:user).joins(:meet).joins(:topic)
-        .select("topics.name as topic_name, users.name as tutor_name, meets.status, meets.date_time")
-        .where(interested: { user_id: @current_user.first.id })
+        query=Meet.joins("INNER JOIN availability_tutors ON availability_tutors.id = meets.availability_tutor_id")
+        .joins("INNER JOIN interesteds ON availability_tutors.id = interesteds.availability_tutor_id")
+        .joins("INNER JOIN users ON availability_tutors.user_id = users.id")
+        .joins("INNER JOIN topics ON availability_tutors.topic_id = topics.id")
+        .select("topics.name AS topic_name,
+               users.name AS tutor_name,
+               meets.status,
+               meets.date_time AS date_time")
+        .where(interesteds: { user_id: @current_user.first.id })
 
         result = query.map do |record|
           {
             topic_name: record.topic_name,
             tutor_name: record.tutor_name,
             meet_status: record.status,
-            meet_date: record.date_time
+            date_time: record.date_time.nil? ? "Indefinida" : record.date_time
           }
         end
         result
-      rescue ActiveRecord::RecordNotFound
-        { error: "Intereted meetings not found" }
-      end
+    rescue ActiveRecord::RecordNotFound
+      { error: "Intereted meetings not found" }
     end